requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
twilio==8.10.0
APScheduler==3.10.4
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Pillow==10.0.1
